Add APIs for start, pause, resume, stop vendors

+++ backend/src/main/java/com/chamod/ticketingbackend/service/TicketPoolService.java
@@ -3,6 +3,6 @@ package com.chamod.ticketingbackend.service;
 public interface TicketPoolService {
 
     void configure(int maxCapacity, int totalTickets);
-    void addTickets(int ticketCount, Long vendorId);
+    void addTickets(int ticketCount, int vendorId);
 
 }
